requirements.txt
django
colorfield
django-phonenumber-field[phonenumbers]
django-countries
Pillow